requirements.txt
numpy
pandas
pyyaml
termcolor
scipy
sklearn
matplotlib
networkx